Move collection dependency database check from importer to worker
